backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
Pillow
web3
supabase
sqlalchemy
psycopg2-binary
